corrección en el borrado de doctores

+++ app/src/main/java/com/me/crudmedico/ui/doctor/presenter/DetailDoctorPresenter.java
@@ -2,6 +2,7 @@ package com.me.crudmedico.ui.doctor.presenter;
 
 import android.content.ContentValues;
 import android.content.Context;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
 import com.me.crudmedico.data.DataBase;
@@ -25,9 +26,16 @@ public class DetailDoctorPresenter implements DetailDoctorContract.Presenter {
     public void deleteDoctor(Doctor doctor) {
         DataBase dataBase = new DataBase(context, "databaseMedicalCrud", null, 1);
         SQLiteDatabase sqLiteDatabase = dataBase.getWritableDatabase();
-        sqLiteDatabase.delete("doctor", "code=" + "'"+doctor.getCode()+"'", null);
+        Cursor cursor = sqLiteDatabase.rawQuery("select * from appointment where codeDoctor='"+doctor.getCode()+"'", null);
+        if(cursor.getCount() > 0){
+            view.get().confirm("No se puede borrar el doctor porque tiene citas en el historial");
+        }
+        else{
+            sqLiteDatabase.delete("doctor", "code=" + "'"+doctor.getCode()+"'", null);
+            view.get().confirm("El doctor fue borrado");
+        }
         sqLiteDatabase.close();
-        view.get().confirm("El doctor fue borrado");
+
     }
 
     @Override
@@ -46,6 +54,14 @@ public class DetailDoctorPresenter implements DetailDoctorContract.Presenter {
         contentValues.put("consultingRoom", doctor.getConsultingRoom());
         contentValues.put("home", doctor.getHome());
         sqLiteDatabase.update("doctor", contentValues, "code="+"'"+doctor.getCode()+"'", null);
+        sqLiteDatabase.close();
+    }
+
+    @Override
+    public void deleteHistory(Doctor doctor) {
+        DataBase dataBase = new DataBase(context, "databaseMedicalCrud", null, 1);
+        SQLiteDatabase sqLiteDatabase = dataBase.getWritableDatabase();
+        sqLiteDatabase.delete("appointment", "codeDoctor=" + "'"+doctor.getCode()+"'", null);
         sqLiteDatabase.close();
     }
 }
